#4 creacion de modulo de infraestructura

+++ com.capitole.domain/src/main/java/com/capitole/price/PriceId.java
@@ -3,5 +3,5 @@ package com.capitole.price;
 import com.capitole.common.StronglyTypedIdentifier;
 
 @StronglyTypedIdentifier
-public record PriceId (int id) {
+public record PriceId (long id) {
 }
